POC for drawing variable loading bars with chars

+++ src/example_load.rs
@@ -6,7 +6,7 @@ use crate::loading_element::LoadingElement;
 
 pub fn sim_load(loading_element: Arc<Mutex<LoadingElement>>, delay: u64) {
     thread::spawn(move || {
-        let max: i32 = loading_element.lock().unwrap().get_max();
+        let max: usize = loading_element.lock().unwrap().get_max();
         for _i in 0..max {
             loading_element.lock().unwrap().update(1);
             thread::sleep(Duration::from_millis(delay));
